start fixing spice and adding helper functions for netlisting
