+++ examples/sky130_inverter/tests/design_inverter/pw1200/ngspice.log
@@ -1,0 +1,6 @@
+
+Note: Compatibility modes selected: ps
+
+
+Circuit: * substrate spice library
+
